REFACTOR test classes to 'Feature' directory
